feat(share2): add share2 page
This page is supposed to be the same as share page, but with code
responsible for "Session" part of library being in different file.
For now, Session code is in different file and properly loaded by
"client" side js code, although there is no functionality that was
provided with share page.